Rearrange some classes in ire
* Move the `TestUtil` class to the test classes
* Move the `GenericMath` class to its own `math` package
* Move message-related classes to the `messages` class
* Move collections and their tests to the `collections` package
* Move `MapperNode` class to the `unarynodes` package
