+++ ire/src/main/scala/hu/bme/mit/ire/collections/BufferMultimap.scala
@@ -1,4 +1,4 @@
-package hu.bme.mit.ire.util
+package hu.bme.mit.ire.collections
 
 import scala.collection.mutable
