+++ ire/src/main/scala/hu/bme/mit/ire/engine/RelationalEngine.scala
@@ -3,10 +3,9 @@ package hu.bme.mit.ire.engine
 import akka.actor._
 import akka.pattern
 import akka.util.Timeout
-import hu.bme.mit.ire._
 import hu.bme.mit.ire.datatypes.Tuple
 import hu.bme.mit.ire.listeners.{AddListener, ChangeListener}
-import hu.bme.mit.ire.messages.{ChangeSet, SizeRequest}
+import hu.bme.mit.ire.messages.{ChangeSet, SizeRequest, Terminator}
 
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
