+++ ire/src/main/scala/hu/bme/mit/ire/nodes/binary/AntiJoinNode.scala
@@ -1,8 +1,7 @@
 package hu.bme.mit.ire.nodes.binary
 
-import hu.bme.mit.ire.SingleForwarder
 import hu.bme.mit.ire.datatypes._
-import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage}
+import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage, SingleForwarder}
 import hu.bme.mit.ire.util.SizeCounter
 
 import scala.collection.mutable
