+++ ire/src/main/scala/hu/bme/mit/ire/nodes/binary/JoinNode.scala
@@ -1,8 +1,7 @@
 package hu.bme.mit.ire.nodes.binary
 
-import hu.bme.mit.ire._
-import hu.bme.mit.ire.datatypes.{Mask, Tuple}
-import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage, TerminatorMessage}
+import hu.bme.mit.ire.datatypes.Mask
+import hu.bme.mit.ire.messages.{ReteMessage, SingleForwarder}
 
 class JoinNode(override val next: (ReteMessage) => Unit,
                override val primaryTupleWidth: Int,
