+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/MapperNode.scala
@@ -1,7 +1,6 @@
 package hu.bme.mit.ire.nodes.unary
 
-import hu.bme.mit.ire.SingleForwarder
-import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage}
+import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage, SingleForwarder}
 
 /**
   * Currently used for selecting elements from a list.
