+++ ire/src/main/scala/hu/bme/mit/ire/nodes/unary/ProjectionNode.scala
@@ -1,8 +1,7 @@
 package hu.bme.mit.ire.nodes.unary
 
-import hu.bme.mit.ire.SingleForwarder
 import hu.bme.mit.ire.datatypes.Tuple
-import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage}
+import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage, SingleForwarder}
 
 abstract class ProjectionImpl(val mask: Vector[Tuple => Any]) extends UnaryNode {
   override def onSizeRequest() = 0
