+++ ire/src/main/scala/hu/bme/mit/ire/stateless/unary/DuplicateEliminationNode.scala
@@ -1,7 +1,6 @@
 package hu.bme.mit.ire.stateless.unary
 
-import hu.bme.mit.ire.SingleForwarder
-import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage}
+import hu.bme.mit.ire.messages.{ChangeSet, ReteMessage, SingleForwarder}
 import hu.bme.mit.ire.nodes.unary.UnaryNode
 
 class DuplicateEliminationNode(override val next: (ReteMessage) => Unit) extends UnaryNode with SingleForwarder {
