+++ ire/src/test/scala/hu/bme/mit/ire/util/TestUtil.scala
